fix: 收口归档任务审查问题
Agent-Logs-Url: https://github.com/Hisoka6602/Zeye.Sorting.Hub/sessions/51715700-9698-4e5e-b270-f2c54294d859

+++ Zeye.Sorting.Hub.Contracts/Models/DataGovernance/ArchiveTaskResponse.cs
@@ -32,12 +32,12 @@ public sealed record ArchiveTaskResponse {
     /// <summary>
     /// 计划候选数量。
     /// </summary>
-    public int PlannedItemCount { get; init; }
+    public long PlannedItemCount { get; init; }
 
     /// <summary>
     /// 已处理数量。
     /// </summary>
-    public int ProcessedItemCount { get; init; }
+    public long ProcessedItemCount { get; init; }
 
     /// <summary>
     /// 发起人。
